requirements.txt
fastapi
pydantic
uvicorn
pandas
sqlalchemy
joblib
psycopg2-binary
scikit-learn
python-dotenv
